+++ cvat-core/src/server-proxy.js
@@ -756,11 +756,7 @@
                         },
                     );
                 } catch (errorData) {
-                    const code = errorData.response ? errorData.response.status : errorData.code;
-                    throw new ServerError(
-                        `Could not get Image Context of the frame for the task ${tid} from the server`,
-                        code,
-                    );
+                    throw generateError(errorData);
                 }
 
                 return response.data;