+++ cvat/apps/engine/tests/test_rest_api.py
@@ -4228,7 +4228,7 @@ class TaskAnnotationAPITestCase(JobAnnotationAPITestCase):
                             "points": [1.0, 2.1, 51, 36.6, 8.0, 11.0],
                             "type": "polygon",
                             "occluded": False,
-                            "outside": False
+                            "outside": True
                         },
                         {
                             "frame": 2,
@@ -4236,7 +4236,7 @@ class TaskAnnotationAPITestCase(JobAnnotationAPITestCase):
                             "points": [1.0, 2.1, 51, 36.6, 14.0, 15.0],
                             "type": "polygon",
                             "occluded": False,
-                            "outside": True,
+                            "outside": False,
                         }
                     ]
                 }]